refactor: delete CTREConfigs.java, move SwerveConstants.java into /swerve
