+++ src/main/java/frc/robot/subsystems/swerve/SwerveConstants.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.subsystems.swerve;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import frc.lib.util.swerveUtil.RevSwerveModuleConstants;
